Refactor - moved files
Moved project files into better named packages, in a better-structured directory structure.
